Change back the argument passing of the dh_ctest_* commands
These commands are meant to be, in their default state, thin
wrappers around their dh_auto_* counterparts, and are meant to be
invoked by dh directly. Passing the dh_auto_* arguments after a "--"
argument doesn't work well for this use case. If we ever add the
ability to change the commands invoked by the dh_ctest_* commands,
we can change the argument passing behavior for that particular
scenario.